refactor/relatorios: cria base modular do workspace
